multi agent v1 with simple test
