--- app/endpoints/multiagents.py
@@ -1,176 +1,0 @@
-from fastapi import APIRouter, Request, Security
-import asyncio
-
-from app.schemas.multiagents import MultiAgentsRequest
-from app.helpers import LanguageModelReranker
-from app.schemas.security import User
-from app.utils.settings import settings
-from app.utils.lifespan import clients, limiter
-from app.utils.security import check_api_key, check_rate_limit
-from app.helpers import SearchManager, InternetManager
-from app.utils.variables import INTERNET_COLLECTION_DISPLAY_ID
-
-router = APIRouter()
-
-explain_choice = {
-    0: "Je ne comprends pas la demande.",
-    1: "Des informations pertinentes ont été trouvées dans la base de données cherchée.",
-    2: "Je n'ai pas trouvé d'informations pertinentes en base de données, mais il me semblait juste de répondre avec mes connaissances générales.",
-    3: "Je n'ai pas trouvé d'informations pertinentes en base de données, et je ne veux pas me mouiller en répondant quelque chose de faux.",
-    4: "La décision d'aller sur internet a été prise pour chercher des informations pertinentes.",
-}
-
-
-def prep_net(body, user):
-    body.collections = [INTERNET_COLLECTION_DISPLAY_ID]
-    search_manager = SearchManager(
-        model_clients=clients.models,
-        search_client=clients.search,
-        internet_manager=InternetManager(
-            model_clients=clients.models,
-            internet_client=clients.internet,
-            default_language_model_id=settings.internet.args.default_language_model,
-            default_embeddings_model_id=settings.internet.args.default_embeddings_model,
-        ),
-    )
-    searches = search_manager.query(
-        collections=body.collections, prompt=body.prompt, method=body.method, k=5, rff_k=5, user=user, score_threshold=body.score_threshold
-    )
-    return searches
-
-
-def get_prompt_teller_multi(question, docs_tmp, choice):
-    """
-    Create a batch of prompts based on docs for the writers maker LLMs.
-    If there is no docs (no context needed), only create one prompt.
-    """
-    prompts = []
-    if choice == 1 or choice == 4:
-        for doc in docs_tmp:
-            prompt_teller = f"""
-Tu es un assistant administratif qui réponds a des questions sur le droit et l'administratif en Français. Tes réponses doit être succinctes et claires. Ne détailles pas inutilement.
-Voilà un contexte : \n{doc}\n
-Voilà une question : {question}
-En ne te basant uniquement sur le contexte donné, réponds à la question avec une réponse de la meilleure qualité possible. 
-- Si le contexte ne te permets pas de répondre à la question, réponds juste "Rien ici", ne dis jamais "le texte ne mentionne pas".
-- Si le contexte donne des éléments de réponse, réponds uniquement a la question et n'inventes rien, donnes même juste quelques éléments de réponse si tu n'arrives pas à répondre totalement avec le contexte. Donnes le nom du texte du contexte dans ta réponse.
-- Si la question n'est pas explicite et renvoie à la conversation en cours, et que tu trouve que le contexte est en lien avec la conversation, réponds juste "Ces informations sont interessantes pour la conversation".
-question : {question}
-réponse ("Rien ici" ou ta réponse): 
-            """
-            prompts.append(prompt_teller)
-    elif choice == 2:
-        for i in range(1):
-            prompt_teller = f"""
-Tu es un assistant administratif qui réponds a des questions sur le droit et l'administratif en Français. Nous sommes en 2024. Tes réponses doit être succinctes et claires. Ne détailles pas inutilement.
-Voilà une demande utilisateur : {question}
-Réponds à cette question comme tu peux. 
-Règles à respecter :
-N'inventes pas de référence.
-Si tu as besoin de plus d'information ou que la question n'est pas claire, dis le a l'utilisateur.
-La réponse doit être la plus courte possible.  Mets en forme ta réponse avec des sauts de lignes. Réponds en Français et part du principe que l'interlocuteur est Français et que ses questions concerne la France.
-Réponse : 
-            """
-            prompts.append(prompt_teller)
-    return prompts
-
-
-def get_completion(model, prompt, user, temperature=0.2, max_tokens=200):
-    response = clients.models[model].chat.completions.create(
-        messages=[{"role": "user", "content": prompt}],
-        model=model,
-        temperature=temperature,
-        max_tokens=max_tokens,
-        stream=False,
-        user=user.name,
-    )
-    result = response.choices[0].message.content
-    return result
-
-
-async def get_completion_async(model, prompt, user, temperature, max_tokens):
-    return get_completion(model, prompt, user, temperature, max_tokens)
-
-
-async def ask_in_parallel(model, prompts, user, max_tokens):
-    tasks = []
-    for prompt in prompts:
-        task = asyncio.create_task(get_completion_async(model, prompt, user, temperature=0.2, max_tokens=max_tokens))
-        tasks.append(task)
-    answers = await asyncio.gather(*tasks)
-    return answers
-
-
-PROMPT_CONCAT = """
-    Tu es un expert pour rédiger les bonnes réponses et expliquer les choses. 
-    Voila plusieurs réponses générées par des agents : {answers}
-    En te basant sur ces réponses, ne gardes que ce qui est utile pour répondre à la question : {prompt}
-    Cites les sources utilisées s'il y en a, mais ne parle jamais des "réponses des agents".
-    Réponds avec une réponse à cette question de la meilleure qualité possible.
-    Si des éléments de réponses sont contradictoire, donnes les quand même à l'utilisateur en expliquant les informations que tu as.
-    Réponds juste à la question, ne dis rien d'autre. Tu dois faire un mélange de ces informations pour ne sortir que l'utile de la meilleure manière possible. Termines ta réponse avec un emoji.
-    Réponse : 
-    """
-
-
-def remove_duplicates(lst):
-    seen = set()
-    return [x for x in lst if not (x in seen or seen.add(x))]
-
-
-@router.post("/multiagents")
-@limiter.limit(settings.default_rate_limit, key_func=lambda request: check_rate_limit(request=request))
-async def multiagents(request: Request, body: MultiAgentsRequest, user: User = Security(check_api_key)):
-    """Multi Agents researcher."""
-
-    reranker = LanguageModelReranker(model=clients.models[body.supervisor_model])
-    client = clients.models[body.supervisor_model]
-    url = f"{client.base_url}multiagents"
-    headers = {"Authorization": f"Bearer {client.api_key}"}
-
-    searches = clients.search.query(
-        prompt=body.prompt,
-        collection_ids=body.collections,
-        method=body.method,
-        k=body.k,
-        rff_k=body.rff_k,
-        score_threshold=body.score_threshold,
-        user=user,
-    )
-    initial_docs = [doc.chunk.content for doc in searches]
-    initial_refs = [doc.chunk.metadata.document_name for doc in searches]
-
-    async def go_multiagents(body, initial_docs, initial_refs, n_retry, max_retry=5, window=5):
-        docs_tmp = initial_docs[n_retry * window : (n_retry + 1) * window]
-        refs_tmp = initial_refs[n_retry * window : (n_retry + 1) * window]
-
-        input = ["(Extrait : " + ref + ") " + doc[:250] + "..." for doc, ref in zip(docs_tmp, refs_tmp)]
-        choice = reranker.create(prompt=body.prompt, input=input, type="choicer")[0].score
-
-        if choice in [0, 3] and n_retry < max_retry:
-            return await go_multiagents(body, initial_docs, initial_refs, n_retry=n_retry + 1, max_retry=5, window=5)
-        elif choice in [1, 2]:
-            pass
-        elif choice == 4 or n_retry >= max_retry:  # else ?
-            searches = prep_net(body, user)
-            docs_tmp = [doc.chunk.content for doc in searches]
-            refs_tmp = [doc.chunk.metadata.document_name for doc in searches]
-        prompts = get_prompt_teller_multi(body.prompt, docs_tmp, choice)
-        answers = await ask_in_parallel(body.writers_model, prompts, user, body.max_tokens_intermediate)
-        prompt = PROMPT_CONCAT.format(prompt=body.prompt, answers=answers)
-        answer = get_completion(body.supervisor_model, prompt, user, temperature=0.2, max_tokens=body.max_tokens)
-        return answer, docs_tmp, refs_tmp, choice, n_retry
-
-    answer, docs_tmp, refs, choice, n_retry = await go_multiagents(body, initial_docs, initial_refs, n_retry=0, max_retry=5, window=5)
-
-    response = {}
-    response["answer"] = answer
-    response["choice"] = choice
-    response["choice_desc"] = explain_choice[choice]
-    response["n_retry"] = n_retry
-    response["sources_refs"] = remove_duplicates(refs)
-    response["sources_content"] = remove_duplicates(docs_tmp)
-    if choice == 2:
-        response["sources_refs"] = ["Trust me bro."]
-        response["sources_content"] = []
-    return response
